Moved test to its own file
